Separate tests from HttpScriptPluginIntegrationSpec that use non-Unicode default charsets
